pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
packages = [{include='src'}]
version = "0.1.0"
description = ""
authors = ["Niilo Kuronen <niilo.kuronen@helsinki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.4"
invoke = "^2.2.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = "^7.7.1"
pylint = "^3.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
